Removed file property from export model

+++ lib/tasks/export.rake
@@ -2,12 +2,11 @@ namespace :export do
   desc "Check catalog exports validity and delete export files when expired"
   task remove_expired: [:environment] do
     # Check all the exports created more than 1 week ago
-    exports = Export.where("created_at < ?", Export.validity.ago).where(:file => true)
+    exports = Export.where("created_at < ?", Export.validity.ago)
 
     # Remove all expired export files but keep the db records
     exports.each do |export|
-      FileUtils.rm_f(export.pathname)
-      export.update(file: false)
+      FileUtils.rm_f(export.pathname) if export.file?
     end
   end
 end
